* renamed some files for consistency:   batchCommands -> fileSystemBatchCommands   batchCommandsMacOnly -> MacOnlyBatchCommands   batchCommandsWinOnly -> WinOnlyBatchCommands   + their test files
