+++ pybatch/conditionalBatchCommands.py
@@ -1,5 +1,5 @@
 from typing import List
-from .batchCommands import *
+from .fileSystemBatchCommands import *
 
 
 class If(PythonBatchCommandBase, essential=True):
